# docs/swagger.yaml
definitions:
...
  api-gateway_pkg_api_v1.CreateInviteLinkResponse:
    properties:
      event_code:
        type: string
    type: object
...
  api-gateway_pkg_api_v1.JoinEventResponse:
    properties:
      event_id:
        type: string
      success:
        description: Успешно/не успешно
        type: boolean
    type: object
  api-gateway_pkg_api_v1.LeaveEventResponse:
    properties:
      success:
        description: Успешно/ не успешно
        type: boolean
    type: object
...
  api-gateway_pkg_api_v1.RemoveParticipantResponse:
    properties:
      success:
        description: Успешно/ не успешно
        type: boolean
    type: object
...
  api-gateway_pkg_response.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
...
  internal_transport_handlers_interaction.CreateLinkInviteDTO:
    properties:
      expires_at:
        type: string
      invite_type:
        type: string
    type: object
  internal_transport_handlers_interaction.JoinEventDTO:
    properties:
      event_code:
        type: string
    type: object
  internal_transport_handlers_participants.ParticipantResponseDTO:
    properties:
      login:
        type: string
      name:
        type: string
      role:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  title: Eventify API Gateway
paths:
...
  /events/{event_id}/invites:
    post:
      consumes:
      - application/json
      description: Тип приглашения (single/multi/unlimited)
      parameters:
      - description: ID мероприятия
        in: path
        name: event_id
        required: true
        type: string
      - description: Настройки инвайта
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_handlers_interaction.CreateLinkInviteDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает event_code
          schema:
            $ref: '#/definitions/api-gateway_pkg_api_v1.CreateInviteLinkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать ссылку приглашение
      tags:
      - interaction
  /events/{event_id}/leave:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID мероприятия
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает статус (Успешно/не успешно)
          schema:
            $ref: '#/definitions/api-gateway_pkg_api_v1.LeaveEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Покинуть ивент
      tags:
      - interaction
  /events/{event_id}/participants:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID мероприятия
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает список участников с именами
          schema:
            items:
              $ref: '#/definitions/internal_transport_handlers_participants.ParticipantResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
      summary: Получить список участников ивента
      tags:
      - participants
  /events/{event_id}/participants/{participant_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID мероприятия
        in: path
        name: event_id
        required: true
        type: string
      - description: ID участника
        in: path
        name: participant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает статус succes
          schema:
            $ref: '#/definitions/api-gateway_pkg_api_v1.RemoveParticipantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: удалить участника из мероприятия
      tags:
      - participants
  /events/join:
    post:
      consumes:
      - application/json
      description: в теле запроса отправляется eventCode
      parameters:
      - description: Код мероприятия
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_handlers_interaction.JoinEventDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает id ивента и статус успешно/безуспешно
          schema:
            $ref: '#/definitions/api-gateway_pkg_api_v1.JoinEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Присоединиться к ивенту
      tags:
      - interaction
...
securityDefinitions:
  BearerAuth:
    description: API для сервиса мероприятий Eventify
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"